// src/modules/user/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  PATIENT = 'PATIENT',
  DOCTOR = 'DOCTOR',
  THERAPIST = 'THERAPIST',
  COUNSELLOR = 'COUNSELLOR',
  LAB_PROFESSIONAL = 'LAB_PROFESSIONAL',
  LAB_STAFF = 'LAB_STAFF',
  LAB_ADMIN = 'LAB_ADMIN',
  HOSPITAL_ADMIN = 'HOSPITAL_ADMIN',
  PHARMACY_ADMIN = 'PHARMACY_ADMIN',
}